Add tests and make compile.

+++ pbj-core/pbj-runtime/src/main/java/com/hedera/pbj/runtime/io/buffer/Bytes.java
@@ -537,7 +537,7 @@ public final class Bytes implements RandomAccessData, Comparable<Bytes> {
     @Override
     public int hashCode() {
         if (hashCode == 0) {
-            hashCode = NonCryptographicHashing.hash64(buffer, start, length);
+            hashCode = (int) NonCryptographicHashing.hash64(buffer, start, length);
         }
         return hashCode;
     }
